_utils/_translation_utils/requirements_notranslate.txt
attrs==19.3.0
certifi==2019.11.28
importlib-metadata==1.5.0
jsonschema==3.2.0
pycurl==7.43.0.5
pyrsistent==0.15.7
python-frontmatter==0.5.0
PyYAML==5.3
six==1.14.0
zipp==3.1.0
